Fix links containing special characters, particularly question mark
For some reason, even when the question mark is percent-encoded, the GitHub
web server returns 404 for the “Why Linton?” page unless it is suffixed with
a trailing slash. (This is probably me misunderstanding something, but after
staring at it for a while I cannot work out what I’m doing wrong!)

Pass on this advice to Linton users too, in the documentation.