Allow linton serve to add “index.html” to URLs (fix #25)
The actual bug was that Linton crashed when trying to serve a directory, but
the least annoying fix is to have the same behaviour as most web servers.

Update the documentation to remove this restriction, but retain the
recommendation that links end in the suffix “index.html” for offline
browsing.